Tag annotations with author and date
Tagging annotations this way makes them much more useful in practise because by its age it's easy to determine its relevance, and because it's attributed it is easy to go to the person who made the annotation for clarification if needed. Git blame can be used for this but can take substantially more effort.