fix ns statements trailing parentheses
The ns statements were breaking the rules of this styleguide concerning the trailing parentheses. I put them each in their own line. I think you shouldn't propose a rule in one part of the document and then break it in the next paragraph, right? Should this also be fixed in the above defn statements?