Threading macros do not have bodies.
The core threading macros `->` and `->>` do not have a clear head/body
distinction, and are not treated or formatted as such in the majority of
extant Clojure code, including the in the implementation of Clojure
itself.

Actual usage of `as->` and `cond->` is less common and more varied, but
seems to support their inclusion in this category.