More nuance on newlines around argument vectors.
The form currently marked “bad” occurs frequently in the implementation
of Clojure itself and many popular libraries.  As it occurs in practice,
is more compact, and doesn’t appear to introduce any readability
problems, I see no reason to advise against it.

The suggested optional new line elision on the other hand leads to more
diff churn when adding docstrings and increases the odds of the
docstring being added incorrectly.  The form occurs somewhat frequently
in practice and thus should be documented as acceptable, but I believe
these caveats should be noted.