Convert to AsciiDoc
The conversion was done in a semi-automatic manner and many headings have to be
manually fixed, as currently they are gibberish. Apart from this the formatting
is in a good state and everything looks well.

This conversion is big step towards publishing an easier to use/navigate
version of the guide and its eventual (potential?) inclusion in clojure.org (which uses
AsciiDoc internally).