Fix false positives in automated spellchecking
By adding an --exclude-file argument, we can add exact lines that
will be matched against to ignore. This prevents us having to add
a word such as "juxt" to the ignored words list when we usually want
to catch that as a misspelling of "just" except in one context.

With all spelling errors fixed and false-positives ignored, the status
of codespell run on each commit can now be trusted.

If any of the lines specified in cs-excludelines.txt need to be modified
in the README, they just need to be adjusted both places.